PIM-4732: Move StorageEvents
