+++ spec/Akeneo/Bundle/StorageUtilsBundle/Doctrine/Common/Saver/BaseSaverSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Akeneo\Bundle\StorageUtilsBundle\Doctrine\Common\Saver;
 
-use Akeneo\Bundle\StorageUtilsBundle\Event\StorageEvents;
+use Akeneo\Component\StorageUtils\StorageEvents;
 use Akeneo\Component\StorageUtils\Saver\SavingOptionsResolverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
